# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Travel Planning API
  version: 0.3.0
  description: >-
    Canonical travel planning API used alongside the MCP server. AI-originated
    mutations become ChangeProposal objects. Approval/apply/rollback operations
    use a separate operator credential. Retry-sensitive writes support
    Idempotency-Key and return Idempotent-Replayed when a stored response is replayed.
servers:
  - url: http://127.0.0.1:8787/v1
    description: Default local development server
security:
  - bearerAuth: []
paths:
...
  /trips:
    get:
      summary: List trips
      responses:
        '200':
          description: Trip summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [trip_id, version, title, start_date, end_date, status, updated_at]
                  properties:
                    trip_id: { type: string, format: uuid }
                    version: { type: integer, minimum: 1 }
                    title: { type: string }
                    start_date: { type: string, format: date }
                    end_date: { type: string, format: date }
                    status: { type: string }
                    updated_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /places/search:
    get:
      summary: Search the configured PlaceProvider
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 25, default: 10 }
      responses:
        '200':
          description: Normalized places plus provider descriptor
          content:
            application/json:
              schema:
                type: object
                required: [provider, places]
                properties:
                  provider: { $ref: '#/components/schemas/ProviderDescriptor' }
                  places:
                    type: array
                    items: { $ref: '../schemas/place.schema.json' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /routes/estimate:
    post:
      summary: Calculate a route through the configured RouteProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_place_id, to_place_id, mode]
              properties:
                from_place_id: { type: string, format: uuid }
                to_place_id: { type: string, format: uuid }
                mode: { enum: [walking, transit, rail, taxi, car, bike] }
      responses:
        '200': { description: Route result with source and calculated_at metadata }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: TRAVEL_API_KEY when configured. Required for non-loopback deployments.
...
  schemas:
    ProviderDescriptor:
      type: object
      required: [id, live, description]
      properties:
        id: { type: string }
        live: { type: boolean }
        description: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request or missing required Idempotency-Key
    Unauthorized:
      description: Missing or invalid Bearer token
...